incometax_project/requirements.txt
# Django and core dependencies
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.4.0
channels==4.0.0
daphne==4.1.2

# Database
dj-database-url==2.2.0
psycopg2-binary==2.9.7

# Task queue
celery==5.3.6
redis==5.0.1

# Production server
gunicorn==21.2.0

# AI and document processing
llama-index-llms-ollama==0.1.3
pandas==2.1.1

# Document processing dependencies
PyMuPDF==1.23.9
camelot-py==0.11.0
openpyxl==3.1.2
xlrd==2.0.1
PyPDF2==3.0.1
pdfplumber==0.10.2

# PDF processing dependencies
opencv-python==4.8.1.78
ghostscript==0.7

# Utilities
python-decouple==3.8
whitenoise==6.6.0

# Monitoring (optional)
flower==2.0.1